resolve comments 4
